[core] Interface and creation of memory pools
Provide interface for the basic memory pools. There'll be some more
functions in future, like allocating from current context (once we have
contexts).

Implement the creation of memory pools. Not tested or tried in any way,
but it should work.

Also provide declaration of the die function that is needed in the
implementation. That one is not implemented yet.