[core] Avoid tiny race condition
It could have happened we entered the while loop around epoll_wait, then
CTRL+C was pressed and then the epoll_wait started, taking a long time
to finish. Now it the CTRL+C will wait to be delivered inside the
epoll_pwait, interrupting it and doing just the last iteration of loop
and terminating fast.