[core] Reset next on plugin copy
When we copy the plugin, we may not keep the old next value. That one
would be dangling pointer and it could lead to strange crashes,
involving call of function outside mapped region of memory (the function
was in unloaded plugin library before). Gdb, quite obviously, wasn't
helpful in providing the name of the function.