[core] Reconnect uplink from fresh loop iteration
If we were to reconnect right away, it could happen the disconnect
plugin callback crashed. That would longjump to the loop and wouldn't
finish the function in uplink. That would never reconnect.

By scheduling the reconnect first (and delaying for later by the loop),
we are sure the attempt will happen.