[master] Compute anomolous buckets
* Fix several errors in the statistics
* Call the statistics on gathered data

There's no real guarantee it computes what it should, but it seems the
high numbers are with the buckets that differ a lot.

We do skip all the all-zero buckets. They'd be rare in practice and the
statistics don't work on them. But we may want to do something about
them later.