[master] Group clients together
Allow clients to be present in arbitrary groups. Process anomalies on
the groups.

Currently, each client is in two groups ‒ one global and one where it is
alone. This will probably have to be revisited (client alone makes only
little sense and makes the statistics grumpy), but let's use this for
testing.