Don't confuse getaddrinfo with complex options
The version on the routers (probably from uclibc) omits several
addresses (the V4 ones) when the options are used, even when we ask for
AI_ALL too. This way, it seems to produce the addresses and match the
server address.

Also, warn if one family of the addresses is missing.