Workaround for connectivity problems after plugin reinit
The problem why socat is not able to reconnect after plugin reinit and
complains something about fopen and missing file (even if no file that
was there before is not missing) is not known. Therefore, as a
workaround, terminate ucollect if the plugin should be reinited and wait
for the watchdog to start it again.

References: #2887.