Fix access-after-release bug
The buffer where the name of plugin was allocated from was then reset,
freeing it. However, until we had a large amount of data in the message,
this wasn't a problem, since that page got reused by the memory pools.
If it gets into a new page (because there's a lot of data already), the
page gets dropped and it segfaults.