flows: Don't loop asking for config
Previously, if the diff data structure requested complete rebuild, the
plugin asked for new config. The config came, but it happened to be the
same ID, so it was thrown away. Then offers of new versions of the data
structure came and it asked to be rebuilt again. So this created a loop
asking for new config.

We reset the config ID when asking for rebuild (and the server should
never have config no. 0). This way it shall think the config is
different and do the actual rebuild.