Don't resend the flow config every two minutes
Due to a bug in 6c2624765227d3a7157ce558ff61bd36ff55c858, we were
resending the config every two minutes. We resend a config when we mark
a client as it asked for a config. However, to resend it, we mark the
client again for the next 2-minute long while, creating a loose loop.

We now have three states:
• We haven't seen the client at all
• We have seen it, and answered the first request for config
• We have seen it and we haven't answered at least one of the requests

We resend it only in the third case, but resending it puts the client
into the second, so the loop is broken.