amihacked: Swap indexes
Use the attack kind as the first index and the date as second. First,
the resulting JSONs are likely to be smaller, second, it makes sense to
do a graph of one kind of attack over time, but it makes little sense to
do a graph for single day comparing pears and apples.